Honor project's repositories when resolving parent coords
Prior to this commit, when the plugin was resolving the parent
project by it's maven coordinates, it did not pass any
repository info to aether.  This meant that aether would only
look in the user's .m2 cache, or on maven central, to try
to find the parent artifact.

This was problematic for projects whose parent project is published
to clojars or to a private maven artifact repository.

This commit grabs the repositories configuration from the child
project and includes it in the call to aether, allowing the
artifacts to be found from other places besides maven central.