Use META-INF to fix broken project.clj path in Lein 2.8.0+
Prior to this commit, in leiningen from 2.0.0-RC1 to 2.7.1, project.clj files were placed at the root of the archive or .jar. Post leiningen 1.8.0-rc, the project.clj file was no longer added to the root of the jar, therefore, calls to project/read failed, as dependencies were not resolved from a parent project's artifact coordinates entry. With this commit, the path for finding a parents project.clj file is changed to the META-INF folder that contains project.clj for leiningen releases post 2.0.0-RC1.