List dependencies in dev profile
We currently list pomegranate and plexus as dependencies and have then
evaluated within leiningen's classpath. Leiningen depends on these tools
itself and consequently always warns that it attempted to load our
version of the tools but didn't because it had already loaded its
versions (when the versions are not the same).

Since we will always have our dependencies evaluated within leiningen's
classpath during runtime and need to depend on the same dependency
resolver it uses, we should only specify our dependencies in our dev
environment.

Note: this removes the dependency on plexus since it seems to be unused
in our requires.